+++ tormenta-core/src/main/scala/com/twitter/tormenta/spout/FlatMappedSchemeSpout.scala
@@ -16,9 +16,8 @@ limitations under the License.
 
 package com.twitter.tormenta.spout
 
-import backtype.storm.topology.IRichSpout
 import com.twitter.tormenta.scheme.Scheme
-import backtype.storm.task.TopologyContext
+import org.apache.storm.task.TopologyContext
 
 /**
  * Spout that performs a flatMap operation on its contained
@@ -27,13 +26,14 @@ import backtype.storm.task.TopologyContext
  */
 
 class FlatMappedSchemeSpout[-T, +U](spout: SchemeSpout[T])(fn: T => TraversableOnce[U])
-    extends SchemeSpout[U] {
+    extends SchemeSpout[U] { self =>
   override def getSpout = spout.getSpout(_.flatMap(fn), callOnOpen)
   override def getSpout[R](transform: Scheme[U] => Scheme[R], f: => TopologyContext => Unit) =
     spout.getSpout(scheme => transform(scheme.flatMap(fn)), f)
+  override def callOnOpen = spout.callOnOpen
 
   override def openHook(f: => TopologyContext => Unit) =
     new FlatMappedSchemeSpout[T, U](spout)(fn) {
-      override def callOnOpen = (c: TopologyContext) => { f(c); spout.callOnOpen(c) }
+      override def callOnOpen = (c: TopologyContext) => { f(c); self.callOnOpen(c) }
     }
 }